Updated to MAME 0.136u1 core
git-svn-id: svn://messdev.no-ip.org/mess@7075 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mame/drivers/cardline.c
@@ -16,7 +16,7 @@
 ***********************************/
 
 
-#include "driver.h"
+#include "emu.h"
 #include "cpu/mcs51/mcs51.h"
 #include "sound/okim6295.h"
 
@@ -31,7 +31,7 @@ static UINT8 *colorram;
 					(colorram[index+offset]&0x80)>>7,\
 					0,0,\
 					x<<3, y<<3,\
-					transparency?transparency:-1);
+					transparency?transparency:(UINT32)-1);
 
 static VIDEO_UPDATE( cardline )
 {
